refactor: revisit Haspara.Accounting module and its submodules
