--- src/Haspara/Accounting/Types.hs
@@ -1,7 +1,0 @@
-module Haspara.Accounting.Types where
-
-import Haspara.Quantity (Quantity)
-import Refined          (NonNegative, Refined)
-
-
-type UnsignedQuantity s = Refined NonNegative (Quantity s)
